Move Harness wrapper to a different file
