+++ research/conditional/train/cc_train.py
@@ -10,7 +10,8 @@ import torch.multiprocessing as mp
 from torch.distributed import init_process_group, destroy_process_group
 
 from lizrd.core import misc
-from lizrd.core.llm import EmbeddingLayer, HarnessLM, Parallel
+from lizrd.core.harness_wrapper import HarnessLM
+from lizrd.core.llm import EmbeddingLayer, Parallel
 from lizrd.support.logging import get_current_logger, get_logger
 from lizrd.support.misc import (
     get_argument_attributes,
